requirements.txt
attrs==21.2.0
click==8.0.1
Flask==2.0.1
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
packaging==20.9
pluggy==0.13.1
protobuf==3.17.3
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
six==1.16.0
toml==0.10.2
Werkzeug==2.0.1
configparser==5.0.2
pyOpenSSL==20.0.1
PyJWT==2.1.0
